+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Lib: add submitEvent to Dom_html and update appropriate function signatures
 * Compiler: complete support for OCaml 4.12
 * Lib: expose API to attached and retrieve js errors to/from ocaml exceptions
+* Lib: intersection observer API fixes
 
 # 3.8.0 (2020-10-21) - London
 ## Features/Changes