+++ CHANGES.md
@@ -6,6 +6,9 @@
 * Lib: expose API to attached and retrieve js errors to/from ocaml exceptions
 * Lib: intersection observer API fixes
 
+## Bug fixes
+* Compiler: fix a segmentation fault when `flat-float-array` mode is disabled.
+
 # 3.8.0 (2020-10-21) - London
 ## Features/Changes
 * compiler, ppx_js, ppx_deriving_json: port PPX's to ppxlib (#1041)